Fix TextAppearanceListener only being applied on last tag
This fixes a bug where the TextAppearance was only applied on the last occurned of a tag. For example for `<tag1>foo</tag1><tag1>bar</tag1>`, the TextAppearance was only applied to `bar`.

This was caused by the same span being reused because of the way BlockStyleListener is implemented. This fixes this issue by simply creating a TextAppearanceSpan for every usage instead, which aligns it with the other listeners.